fix: newCakeStakersTotalEstimateRewardUSD
